refactor: RecentNotificationsWidgetをDashboardPageから分離
最近の通知セクション（49行）を独立コンポーネントに抽出。
notificationNameMap/getNotificationTextもウィジェットに内包。
DashboardPage: 350→288行（62行削減）。
